Release resource left.
Create resource on resource manager works. Only on the graphic side,
both textures and modules left. Sound too.

Release a resource is left, but to do this I need to loop through the
resources to find the resource(type, id etc) and get the resource name.
Check all other resources in the same type if theres any other pointing
to the same object, if so, delete only the id, e.g. not model type.