Class splitt
Splitted the modelconverter class so now there is three classes, one
that reads .tx, the converter and one that reads the .btx binary file. I
have also implemented joints and weights in the new exporter class. To
do is: Clean up code and merge it with the havenborough project. Then
testing.