AH-244 Idea of how it should work.
An good idea of how it should work, much like how graphic engine handles model and modelinstances.
Load low poly model, have it stored in physics with a name. When level want to create a hull check this list to see if it is loaded and copy triangles to the hull. Create body and return the bodyhandle to level so it can manipulate the bodies.