AH-407 Mana regeneration, mana cost and mana checks!
There is now mana regeneration in the game. More mana is regenerated when moving at full speed.
Spells now have mana cost. There is also a check that requires the currentmana to be more than the mana cost to be able to cast a spell.

However there is something fishy about how spelldefinitions, can't really explain it... :(